before templating
